v1.5b 增加了黑鸟等四个英雄，及一些微调 (#11)
+v1.5b(2018.4.13) committed by zmcmcc
+1. Added new hero Outworld Destroyer, Naga Siren, Terrorblade and Templar Assassin.
+2. Now if bots didn't choose a safe lane support, the other support will buy the courier.
+3. Some heros will now purchase Aeon Disk.
+4. Fixed the bug that bots will pick duplicated heroes in Captain mode.
+5. Slightly adjust the ability usage of Centaur, Slardar, Ursa etc. so that they are less likely to miss their stuns.
+6. Rewrite the ability usage of skeleton king. Althogh he still can't use his third ability correctly, he has better ability use now.
+7. Rewrite the ability usage of pugna. Now his ult still has some problems(Similar to SB and ET) but has better ability use now.
+8. Rewrite the ability usage of tidehunter. Now he can use gush correctly.

+++ ability_item_usage_earth_spirit.lua
@@ -367,6 +367,13 @@ Consider[2]=function()
 	--------------------------------------
 	-- Global high-priorty usage
 	--------------------------------------
+	--Check if stuck on cliff.
+	if not IsLocationPassable(npcBot:GetLocation()) then
+		local loc = npcBot:GetXUnitsTowardsLocation( location, nUnitCR );
+		return BOT_MODE_DESIRE_HIGH,loc, false;
+	end
+
+
 	--Try to kill enemy hero
 	if(npcBot:GetActiveMode() ~= BOT_MODE_RETREAT ) 
 	then
